The marshalOID () and oidToString () functions use the string functions Trim, Split, Join, which work internally with an array of strings. Also, to convert the oid string to an array of numbers and in the opposite direction, an additional array is created inside both marshalOID () and oidToString (), which causes at least twice the load on the GC.
In my case, with parallel polling of 100k + devices, this change in these functions improved performance and CPU load by 10% -15%. The average CPU load dropped from 80% to 65% after. In addition, the total polling time has noticeably decreased.

The second patch (reverted due to broken ci check) concerns the logging function logger.Printf(). Even with logging disabled, the Printf() system call is still called, which, according to pprof, also consumes a significant portion of the CPU time.
I don't know what to do in this case, so I just commented them out in problem areas. Logically, the logPrintf () function should be called here, but in these cases it is not available for call.