@TimRots 2021-04-05
Thank you for another great PR @it1804.

Perhaps it's better to split-off the second patch you mention into a separate PR?
I btw do see some overlap with #313, please feel free to join in the conversation there if you think the issue you mention can be fixed in parallel. Otherwise opening a separate issue to untangle things would be nice to have.