@it1804 2021-04-05
> Perhaps it's better to split-off the second patch you mention into a separate PR? As I do see some overlap with #313 please feel free to join in the conversation there if you think the issue you mention can be fixed in parallel.
> Otherwise opening a separate issue to untangle things would be nice to have.

Yes, I will join that discussion. I think that both logging problems can be solved in one place. The second patch shouldn't have been included in this pr.