@it1804 2021-04-07
My tests are simple. The functions are called in a loop with one million iterations.
In the case of oidToString, I additionally tested one more implementation using a byte buffer, it is close to my result, but slightly worse.
For the oidToIntArray function, I took the marshalOID part, in which the string is directly converted into an array of numbers.
You need to run it locally on your computer, The Go Playground does not allow you to run the program with a long timeout. 

[Benchmark oidToIntArray](https://play.golang.org/p/B_TTWDZlDl_m)
[Benchmark oidToString](https://play.golang.org/p/MQvzWmRVHi6)