@TimRots 2021-04-07
> My tests are simple. The functions are called in a loop with one million iterations.
> In the case of oidToString, I additionally tested another implementation using a byte buffer, it is close to my result, but a little slower.
> For the oidToIntArray function, I took the marshalOID part, in which the string is directly converted into an array of numbers.
> You need to run it locally on your computer, The Go Playground does not allow you to run the program with a long timeout.
> 
> [Benchmark oidToIntArray](https://play.golang.org/p/cvujPO8-BEj)
> [Benchmark oidToString](https://play.golang.org/p/MQvzWmRVHi6)

Thank you for updating the PR and providing these tests @it1804. The performance improvements are impressive :)

```
marshalOID_Partial_Orig time: 3.713267932s
marshalOID_Partial_New time: 1.704001884s

oidToStringOrig time: 2.788383612s
oidToStringWithStringBuilder time: 182.199084ms
oidToStringWithBytesBuffer time: 227.554092ms
```

@SuperQ Could you give this PR a review please?