@it1804 2021-04-07
I am working on a couple more improvements, but I'm not sure of an equally significant improvement yet. I think if everything works out, I will make a new PR