@it1804 2021-04-08
@TimRots 
Could you run another test. I added calculating and setting the string buffer size to the oidToStringWithStringBuilderGrowed () function. It also greatly improves the performance in my production environment under high load. To calculate the length of the string oid from the array, an intermediate function was added parseOidLength

[Benchmark oidToString with string Growed](https://play.golang.org/p/O8RgVYz3WCp)