@it1804 2021-04-12
> Can you add micro-benchmarks, so we can see the change in allocs? I'd have expected all of this to be on the stack.

[bench_oid_to_int_array_test.go](https://play.golang.org/p/qbhCZ1PqU5D)
go test -bench=. -benchmem -benchtime=100000x bench_oid_to_int_array_test.go
Benchmark_marshalOID_Orig-24              100000              2296 ns/op             408 B/op          5 allocs/op
Benchmark_marshalOID_New-24               100000              1324 ns/op             208 B/op          3 allocs/op
PASS
ok      command-line-arguments  0.385s

[bench_oid_to_string_test.go](https://play.golang.org/p/8AUJofpysxj)
go test -bench=. -benchmem -benchtime=100000x bench_oid_to_string_test.go
Benchmark_oidToString_Orig-24                             100000              2444 ns/op             336 B/op         11 allocs/op
Benchmark_oidToString_WithStringBuilderGrowed-24          100000              1739 ns/op             144 B/op         10 allocs/op
Benchmark_oidToString_WithStringBuilder-24                100000              1962 ns/op             320 B/op         14 allocs/op
Benchmark_oidToString_WithBytesBuffer-24                  100000              2525 ns/op             352 B/op         12 allocs/op
PASS
ok      command-line-arguments  0.891s