@efronlicht 2021-04-13
This is a big issue for my employer. We were able to come up with a solution that involves only one allocation.

I'm not sure if I can legally just give you our code, but what I _can_ do is give you a hint:

the algorithm you want looks something like this:

- count the size in bytes of our output. this should be the base-10 length of each section, +1 for each section.
(i.e, `[]int{23, 3}` -> `".23.3"` -> `"." + "23" + "." + "3"`
- allocate a buffer of `[]byte` exactly once.
- write to the buffer. `strconv` in the standard library should give you some hints on how to do this efficiently.
- convert the buffer to a string. this will do an additional allocation unless you use `unsafe`.

if you do it right, you should have only a single allocation, and it should run significantly faster. good luck!