@TimRots 2021-04-14
> the algorithm you want looks something like this:
> 
>     * count the size in bytes of our output. this should be the base-10 length of each section, +1 for each section. this will involve one iteration through the list.
>       (i.e, `[]int{23, 3}` -> `".23.3"` -> `"." + "23" + "." + "3"`
> 
>     * allocate a buffer of `[]byte` with that size exactly once.
> 
>     * write the periods and base-10 formatted integers into the buffer. `strconv` in the standard library should give you some hints on how to do this efficiently. you should only need to iterate through one more time.
> 
>     * convert the buffer to a string. this will cost an additional allocation unless you use `unsafe`, but you can guarantee the buffer won't escape, so `unsafe` should be fine.
> 
> 
> if you do it right, you should have only a single allocation, and it should run significantly faster. good luck!

As much as I'm liking the described algorithm (& initiative of joining the conversation) @efronlicht. 'Righting' this at the cost of introducing ``unsafe`` in gosnmp is imho not preferred.
It's just too easy to introduce bugs which do not outweigh the benefit of this performance optimization.

If we are able to find a solution without ``unsafe`` that would be much welcomed however!