@it1804 2021-04-14
Without using unsafe in the oidToString () function, if this unsafe is removed, when using the byte array [] there will be 2 allocations in the test, but judging by my tests, the speed will not drop much. Fortunately, the size of the pre-allocated buffer is easy to calculate here. If you do not calculate the buffer size in advance, performance degradation will occur. In terms of speed, compared to my original StringBuilder and BytesBuffer, this is a big success. Thanks @efronlicht for the tip.

In another function that converts the oid string to asn1 byte string - marshalOID (), I also came up with an algorithm that avoids intermediate storage in [] int64, which will eliminate the oidToIntArray () function I wanted to add, and possibly marshalOID () itself instead of which I want to make marshalObjectIdentifier () the main and only one in converting oid string to asn1 byte string. I have not done all the tests yet, but there will be 2 allocations per operation beforehand if everything works out. In the worst case, three. I have not found an easy way to calculate the buffer size in advance without a double pass, because in base128 encoding, the size in bytes can range from 1 to 5 bytes.

In the next few days, I will update the commits with the new changes.