@it1804 2021-04-25
> 
> Testing the updated PR yields the following results:
> 
> ```
> BenchmarkParseObjectIdentifierAsByteArray-4   	  400000	       181.1 ns/op	       0 B/op	       0 allocs/op
> BenchmarkParseObjectIdentifierAsString-4      	  400000	       206.6 ns/op	      48 B/op	       1 allocs/op
> BenchmarkMarshalObjectIdentifier-4            	  400000	       122.6 ns/op	       0 B/op	       0 allocs/op
> BenchmarkSendOneRequest-4                     	  400000	       38394 ns/op	    2520 B/op	      83 allocs/op
> ```
> 
> vs current 1.31.1
> 
> ```
> BenchmarkOidToString-4      	  400000	       176.3 ns/op	     112 B/op	       2 allocs/op
> BenchmarkSendOneRequest-4   	  400000	       39293 ns/op	    3256 B/op	      90 allocs/op
> ```
> 
> while using `go test -bench=. -benchmem -benchtime=400000x *.go`
> 
> @SuperQ @brian-brazil Could we consider having this, based on the added benchmark info?

я выпустил новую версию патча, в котором полностью перешел на использование массива байт вместо bytes.Buffer.
Этот результат понравился мне больше.
Для сравнения go test -bench=. -benchmem -benchtime=400000x *.go:

Current master
`BenchmarkSendOneRequest-24        400000             80557 ns/op            3259 B/op         90 allocs/op
`

Previous patch:
`BenchmarkSendOneRequest-24                                400000             72618 ns/op            2522 B/op         83 allocs/op
`

New patch:
`BenchmarkSendOneRequest-24        400000             70879 ns/op            2114 B/op         79 allocs/op`

I also moved the buffer allocation for marshalLength () and parseObjectIdentifier () back inside these functions. In fact, in the full test, the increase in speed turned out to be small, plus such a code did not seem very nice.
But as it turned out, using a byte array everywhere turned out to be more efficient. 

PS:
And there is an added bonus, as I said, the registration code in the current implementation is the biggest performance issue. I have rewritten the code responsible for this, but have not fully tested it yet. Does it make sense to make these changes to this PR as well?

Test with disabled logging
`BenchmarkSendOneRequest-24                                400000             55143 ns/op            1360 B/op         39 allocs/op`